Deeply modified the Ast relations - Working build
So it's not very beautiful but I needed it for Join.
I added types in the Ast to consider relations as queries. The
parser now parses rels into a Product (r, rels) chain insted of a
list.

Also, I deprecated the pretty print of show_query and others because
it was broken by my modifications and ppx_deriving does so automatically
(and I find it sufficient for debugging). But feel free to rewrite it
(as pprint_query, etc. in order not to mistake it for ppx_deriving funs !)